Updated truth table generation function

Former-commit-id: 562093de412096be7df88d3543494bb9b813b5e7

+++ gatelevel_readers/read_verilog/src/read_nl/generate_simple_netlist.cpp
@@ -351,7 +351,7 @@ void simpleTruthTable(std::string tr, std::string w, std::vector<std::vector<uns
 {
     unsigned long long width = veriValue(w);
     unsigned long long table = veriValue(tr);
-    unsigned p = 1;
+    unsigned long long p = 1;
     for (int i = 0; i < 64; ++i)
     {
         if (p & table)
